Create SshTestBase
