+++ src/test/java/io/hyperfoil/tools/qdup/SshContainerTestBase.java
@@ -40,86 +40,15 @@ import static org.junit.Assert.assertTrue;
 /**
  * This base test class provides and way to run the qDup scripts inside a container.
  */
-public class SshContainerTestBase {
+public class SshContainerTestBase extends SshTestBase {
 
-
-    protected TmpDir tmpDir;
-
-    @Before
-    public void setup(){
-        tmpDir = TmpDir.instance();
-    }
-
-    @After
-    public void cleanUp(){
-        //tmpDir.removeDir();
-        tmpDir = null;
-    }
     private static GenericContainer container;
 
     private static final ScheduledThreadPoolExecutor SCHEDULED_THREAD_POOL_EXECUTOR = new ScheduledThreadPoolExecutor(2);
 
     //added to simplify exchanging files with test container that uses identity
     public Local getLocal(){
-        return new Local(getBuilder().buildConfig(Parser.getInstance()));        
-    }
-    public RunConfigBuilder getBuilder(){
-        return getBuilder("qdup");
-    }
-    public RunConfigBuilder getBuilder(String name){
-        RunConfigBuilder builder = new RunConfigBuilder();
-        builder.setIdentity(getIdentity());
-
-        setIdentityFilePerms(getPath("keys"), getKeyDirPerms());
-        setIdentityFilePerms(getPath("keys/"+name), getPrivKeyPerms());
-
-        //set perms
-        return builder;
-    }
-
-    private static void setIdentityFilePerms(Path identityFilePath, Set<PosixFilePermission> perms){
-        try {
-            Files.setPosixFilePermissions(identityFilePath, perms);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private static Set<PosixFilePermission> getKeyDirPerms(){
-        Set<PosixFilePermission> perms = new HashSet<>();
-        //add owners permission
-        perms.add(PosixFilePermission.OWNER_READ);
-        perms.add(PosixFilePermission.OWNER_WRITE);
-        perms.add(PosixFilePermission.OWNER_EXECUTE);
-        perms.add(PosixFilePermission.GROUP_READ);
-        perms.add(PosixFilePermission.GROUP_WRITE);
-        perms.add(PosixFilePermission.GROUP_EXECUTE);
-        perms.add(PosixFilePermission.OTHERS_READ);
-        perms.add(PosixFilePermission.OTHERS_EXECUTE);
-
-        return perms;
-    }
-
-    private static Set<PosixFilePermission> getPrivKeyPerms(){
-        Set<PosixFilePermission> perms = new HashSet<>();
-        //add owners permission
-        perms.add(PosixFilePermission.OWNER_READ);
-        perms.add(PosixFilePermission.OWNER_WRITE);
-
-        return perms;
-    }
-
-
-    public String getIdentity() {
-        return getPath("keys/qdup").toFile().getPath();
-    }
-
-    public static Path getPath(String subDir){
-        return  Paths.get(
-                SshContainerTestBase.class.getProtectionDomain().getCodeSource().getLocation().getPath()
-        ).resolve(
-                Paths.get(subDir)
-        );
+        return new Local(getBuilder().buildConfig(Parser.getInstance()));
     }
 
     public static void restartContainer(){
@@ -299,14 +228,6 @@ public class SshContainerTestBase {
     }
     public static Host getPasswordHost(){
         return new Host(host.getUserName(),host.getHostName(),"password",host.getPort());
-    }
-
-
-
-    public static InputStream stream(String...input){
-        return new ByteArrayInputStream(
-                String.join("\n", Arrays.asList(input)).getBytes()
-        );
     }
 
     protected static class TmpDir{
